refactor: Rename globalstyle
